Fix: Trying to create a new ROM from context menu crashes shex.
When the ROM_buffer ctor is called with new_file == true, it doesn't call analyze() so the function pointers for pc_to_snes etc never get populated.
This crashes shex.
The fix is just to let the buffer be analyzed on creation of new file, even thought we get "UNRECOGNIZED ROM" in the debug log, this is better than the program straight up crashing.