Feature: Let shex be notified of external file changes on disk
This patch adds a feature to shex that lets it get notified about eventual file changes (e.g. modifications, renames or deletions) about the files that it has currently opened in the editor.

Once a file is closed from the editor, it gets removed from the list of files that shex wants to get notified about.

If a notification arrives one of two things can currently happen:
- File was renamed or deleted, if this happens, shex will just ask if you want to close the tab which was populated with that file. If you say yes, the tab will get closed, if you say no, the tab will remain and upon saving a new file with the old name will be created.
- File was modified, if this happens, shex will ask if you want to reload the file. If you reload the file you'll lose all changes and the view will refresh and show the new edits. If you don't reload the file, you keep the changes and upon saving the modified file will get overwritten. 

To be added: instead of losing the changes upon reloading the modified file, open a diff view that lets you choose what to keep and what to toss away.