Feature: Implement diff view for notifications on file changes.
This patch implements a git-like "diff view" that lets you select which changes to keep and which to toss away when a file opened in the editor is changed on disk. 

It is a WIP implementation for now, but it appears to work just fine for all the things I've tested it with.

It enables you to accept/reject each change one-by-one or accept/reject all the remaining ones at once.

Currently when rejecting a change, shex will re-write the old bytes in the file on disk, effectively restoring them to what they were before the other program modified them.

This window obviously doesn't open when a file is removed/renamed from disk.