Rework the diff view
This patch is a huge rework to how the diff view is handled.

Responding yes to a message box telling you that a change has happened will automatically move the tab to the one that received the signal if it's not the current one.

Instead of opening a new window, now the diff view opens to the side of the interested editor, with a table showing the diffs and controls on the bottom. You can click on any cell on the table and you'll be moved to the corresponding diff.

When multiple diffs are triggered to the same file the table will just reload to show the updated state of the file.

When a diff of another file that's open in shex is triggered and there's already a message box open, the alert will be stalled until the current alert is answered.

Multiple diff views can be opened for different files at the same time in different tabs.

Once all diffs are cleared out, the view will automatically close.